pyproject.toml
[tool.poetry]
name = "insta-mine"
version = "0.1.0"
description = "A tool to fetch Instagram user details using their username or Instagram URL. This project helps you gather basic information about an Instagram profile quickly and efficiently."
authors = ["RenesisTech"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"
black = "^24.4.2"
instaloader = "^4.11"
pytest = "^8.2.2"
 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
